refactor: useTimers 를 이용하여 UpdateTimeDivider 내부 함수 리팩토링
